+++ cryptographic_estimators/MAYOEstimator/mayo_problem.py
@@ -17,9 +17,8 @@
 
 
 from ..base_problem import BaseProblem
-from ..MQEstimator.mq_helper import ngates
 from .mayo_constants import *
-from ..helper import is_prime_power
+from ..helper import is_prime_power, ngates
 from math import log2, ceil
