+++ cryptographic_estimators/MQEstimator/__init__.py
@@ -1,7 +1,8 @@
 from .degree_of_regularity import generic_system, regular_system, semi_regular_system, quadratic_system
 from .mq_algorithm import MQAlgorithm
 from .mq_estimator import MQEstimator
-from .mq_helper import ngates, nmonomials_of_degree, nmonomials_up_to_degree, sum_of_binomial_coefficients
+from .mq_helper import nmonomials_of_degree, nmonomials_up_to_degree, sum_of_binomial_coefficients
+from ..helper import ngates
 from .mq_problem import MQProblem
 from .witness_degree import semi_regular_system, quadratic_system
 from .MQAlgorithms import Bjorklund, BooleanSolveFXL, CGMTA, Crossbred, DinurFirst, DinurSecond, ExhaustiveSearch, F5, HybridF5, KPG, Lokshtanov, MHT
