+++ cryptographic_estimators/MQEstimator/mq_problem.py
@@ -17,13 +17,12 @@
 
 
 from cryptographic_estimators.base_problem import BaseProblem
-from cryptographic_estimators.MQEstimator.mq_helper import ngates
 from cryptographic_estimators.MQEstimator.mq_constants import (
     MQ_NUMBER_VARIABLES,
     MQ_NUMBER_POLYNOMIALS,
     MQ_FIELD_SIZE,
 )
-from cryptographic_estimators.helper import is_prime_power
+from cryptographic_estimators.helper import is_prime_power, ngates
 from math import log2, ceil
